// backend/src/types.ts
export interface Context {
  req: Request;
  res: Response;
}

export interface AuthenticatedContext extends Context {
  user: IUserDocument;
  token: string;
}

export type CreateContextFn = ({
  req,
  res,
}: trpcExpress.CreateExpressContextOptions) => Context;

export interface IUser {
  _id: Types.ObjectId;
  email: string;
  password: string;
  name: string;
  authProvider?: "local" | "google";
  googleId?: string;
  avatar?: string;
  refreshToken?: string;
  apiKey?: string;
  apiKeyExpiresAt?: Date;
  isDeleted: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUserDocument extends IUser, Document {
  _id: Types.ObjectId;
  isPasswordCorrect(password: string): Promise<boolean>;
  generateAccessToken(): Promise<string>;
  generateRefreshToken(): Promise<string>;
  generateApiKey(expiryDays?: number): string;
}
// ...
export interface JWTPayload {
  _id: string;
  email?: string;
  name?: string;
  exp?: number;
  iat?: number;
}
// ...
export interface UserSignupInput {
  email: string;
  password: string;
  name: string;
  otp: string;
}

export interface UserLoginInput {
  email: string;
  password: string;
}

export interface ChangePasswordInput {
  oldPassword: string;
  newPassword: string;
}

export interface UpdateAccountInput {
  name?: string;
  email?: string;
  otp?: string;
}

export interface DeleteAccountInput {
  password: string;
}

export interface RefreshTokenInput {
  refreshToken?: string;
}

export interface GoogleAuthUrlInput {
  mode?: "register" | "login";
  state?: string;
  redirectUri?: string;
}

export interface GoogleAuthCodeInput {
  code: string;
  redirectUri?: string;
}

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
}

export interface UserPublic {
  _id: string;
  email: string;
  name: string;
  authProvider?: "local" | "google";
  avatar?: string;
  apiKey?: string;
  createdAt: Date;
  updatedAt: Date;
}